requirements.txt
argcomplete
openidc-client
python-bugzilla
rpkg >= 1.65
six
